media-source/mediasource-duration.html

<!DOCTYPE html>
<html>
    <head>
        <title>MediaSource.duration &amp; HTMLMediaElement.duration test cases.</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="mediasource-util.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
          mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
          {
              mediaSource.duration = 2;
              assert_false(sourceBuffer.updating,
                'No SourceBuffer update when duration is set');
              test.done();
          }, 'MediaSource.duration: setting the duration does not trigger any SourceBuffer update');

          mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
          {
              assert_less_than(segmentInfo.duration, 60, 'Sufficient test media duration');
              sourceBuffer.appendBuffer(mediaData);
              test.expectEvent(sourceBuffer, 'updateend',
                'Media data appended to the SourceBuffer');
              test.waitForExpectedEvents(function()
              {
                  mediaSource.duration = 60;
                  assert_false(sourceBuffer.updating,
                    'No SourceBuffer update when duration is increased');
                  test.done();
              });
          }, 'MediaSource.duration: Increasing the duration does not trigger any SourceBuffer update');

          mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
          {
              assert_greater_than(segmentInfo.duration, 2, 'Sufficient test media duration');
              mediaElement.play();
              sourceBuffer.appendBuffer(mediaData);
              test.expectEvent(sourceBuffer, 'updateend',
                'Media data appended to the SourceBuffer');
              test.waitForExpectedEvents(function()
              {
                  mediaSource.duration = 60;
                  assert_false(sourceBuffer.updating,
                    'No SourceBuffer update when duration is truncated');
                  test.done();
              });
          }, 'MediaSource.duration: Increasing the duration during media playback does not trigger any SourceBuffer update');

          mediasource_testafterdataloaded(function(test, mediaElement, mediaSource, segmentInfo, sourceBuffer, mediaData)
          {
              assert_greater_than(segmentInfo.duration, 2, 'Sufficient test media duration');

              var fullDuration = segmentInfo.duration;
              var newDuration = fullDuration + 0.5;

              var expectedDurationChangeEventCount = 1;
              var durationchangeEventCounter = 0;
              var durationchangeEventHandler = test.step_func(function(event)
              {
                  assert_equals(mediaElement.duration, newDuration, 'mediaElement newDuration');
                  assert_equals(mediaSource.duration, newDuration, 'mediaSource newDuration');
                  durationchangeEventCounter++;
              });

              mediaElement.play();

              // Append all the segments
              test.expectEvent(sourceBuffer, 'updateend', 'sourceBuffer');
              test.expectEvent(mediaElement, 'playing', 'Playing triggered');
              sourceBuffer.appendBuffer(mediaData);

              test.waitForExpectedEvents(function()
              {
                  assert_equals(mediaElement.duration, fullDuration, 'mediaElement fullDuration');
                  assert_equals(mediaSource.duration, fullDuration, 'mediaSource fullDuration');
                  assert_less_than(mediaElement.currentTime, newDuration / 2, 'mediaElement currentTime');

                  // Media load also fires 'durationchange' event, so only start counting them now.
                  mediaElement.addEventListener('durationchange', durationchangeEventHandler);

                  assert_false(sourceBuffer.updating, "updating");

                  // Truncate duration. This should result in one 'durationchange' event fired.
                  mediaSource.duration = newDuration;
                  assert_equals(mediaSource.duration, newDuration);

                  // Skip media to "near the end" to make test take less time
                  mediaElement.currentTime = mediaSource.duration - 0.5;

                  // Set duration again to make sure it does not trigger another 'durationchange' event.
                  mediaSource.duration = newDuration;

                  // Mark endOfStream so that playback can reach 'ended' at the new duration.
                  test.expectEvent(mediaSource, 'sourceended', 'endOfStream acknowledged');
                  mediaSource.endOfStream();

                  // endOfStream can change duration slightly.
                  // Allow for one more 'durationchange' event only in this case.
                  var currentDuration = mediaSource.duration;
                  if (currentDuration != newDuration) {
                      newDuration = currentDuration;
                      ++expectedDurationChangeEventCount;
                  }


                  // Allow media to play to end while counting 'durationchange' events.
                  test.expectEvent(mediaElement, 'ended', 'Playback ended');
                  test.waitForExpectedEvents(function()
                  {
                      mediaElement.removeEventListener('durationchange', durationchangeEventHandler);
                      assert_equals(durationchangeEventCounter, expectedDurationChangeEventCount, 'durationchanges');
                      test.done();
                  });
              });
          }, 'MediaSource.duration: Setting the same duration multiple times does not fire duplicate durationchange events');
        </script>
    </body>
</html>
